requirements.txt
streamlit>=1.32.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0
langchain-chroma>=0.1.0
langgraph>=0.1.0
chromadb>=0.5.0
pypdf>=4.0.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
langchain-tavily>=0.1.0
